Fix five console warnings
- The onClick to open the Drawer was in the MenuIcon and not the
IconButton, this was causing a warning as the onClick should be in the
button. The onClick property was moved to the button. Also, the Drawer
component, that previously was inside the IconButton was moved to be
insed the AppBar component. Without this the Drawer was not closing
after moving the onClick property.
- The key property on the index pages of ApresentacaoPoster,
MesasRedondas, Minicursos and SeminariosTematicos was not in the
direct children. The react fragment that was wrapping the li and the
hr was removed, the hr has moved to the li and now the li, that already
had the key, is the unique children of the map. Also because of that the
css of hr changed a bit.
- In SeminariosTematicos show there were children of maps without keys.
The ul was added wrapping the map and now the added li contain the
needed keys property.
- In SeminariosTematicos show there were elements with class and not
ClassName and there were a typo in one key. This was corrected.
- The React.StrictMode was removed from src/index.js, as the Drawer
component and links passed were not compatible with React.StrictMode.
- Update app version to 1.0.0.